+++ src/createGlobalContext/index.ts
@@ -1,6 +1,4 @@
 export { createGlobalContext } from './createGlobalContext';
-export type {
-  CreateGlobalContextReducers,
-  GlobalContextValue,
-} from './createGlobalContext';
+
+export type { GlobalContextReducers, GlobalContextValue } from './type';
 export type { GlobalContextProviderProps } from './createProvider';
